+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Expected: February 2024
 
 ### New features
 
+* Replaced creator attributes with a configured solution
 * New: [show creator paths associated to service instance](https://github.com/clicon/clixon-controller/issues/90)
   * See https://clixon-controller-docs.readthedocs.io/en/latest/cli.html#creators
 * New CLI commands: